+++ README.md
@@ -1,2 +1,5 @@
-# func-anonymous
-Repository with ready to use functions for K-anonymity and L-diversity
+# open-data
+Repository with ready to use functions for :
+> K-anonymity
+> L-diversity
+> Geopandas data visualization
